pyproject.toml
[tool.poetry]
name = "henshin"
version = "0.1.0"
description = ""
authors = ["Rizky Maulana Nugraha <lana.pcfre@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
alembic = "1.11.1"
anyio = "3.7.1"
apache-airflow = "2.6.3"
apache-airflow-providers-common-sql = "1.5.2"
apache-airflow-providers-ftp = "3.4.2"
apache-airflow-providers-http = "4.4.2"
apache-airflow-providers-imap = "3.2.2"
apache-airflow-providers-sqlite = "3.4.2"
apispec = "5.2.2"
argcomplete = "3.1.1"
asgiref = "3.7.2"
async-timeout = "4.0.2"
attrs = "23.1.0"
babel = "2.12.1"
blinker = "1.6.2"
cachelib = "0.9.0"
cattrs = "23.1.2"
certifi = "2023.5.7"
cffi = "1.15.1"
charset-normalizer = "3.1.0"
click = "8.1.4"
clickclick = "20.10.2"
colorama = "0.4.6"
colorlog = "4.8.0"
configupdater = "3.1.1"
connexion = "2.14.2"
cron-descriptor = "1.4.0"
croniter = "1.4.1"
cryptography = "40.0.2"
deprecated = "1.2.14"
dill = "0.3.1.1"
dnspython = "2.3.0"
docutils = "0.20.1"
email-validator = "1.3.1"
exceptiongroup = "1.1.2"
flask = "2.2.5"
flask-appbuilder = "4.3.1"
flask-babel = "2.0.0"
flask-caching = "2.0.2"
flask-jwt-extended = "4.5.2"
flask-limiter = "3.3.1"
flask-login = "0.6.2"
flask-session = "0.5.0"
flask-sqlalchemy = "2.5.1"
flask-wtf = "1.1.1"
frozenlist = "1.3.3"
google-re2 = "1.0"
graphviz = "0.20.1"
gunicorn = "20.1.0"
h11 = "0.14.0"
httpcore = "0.16.3"
httpx = "0.23.3"
idna = "3.4"
importlib-resources = "5.12.0"
inflection = "0.5.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
jsonschema = "4.18.0"
jsonschema-specifications = "2023.6.1"
lazy-object-proxy = "1.9.0"
limits = "3.5.0"
linkify-it-py = "2.0.2"
lockfile = "0.12.2"
mako = "1.2.4"
markdown = "3.4.3"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.19.0"
marshmallow-enum = "1.5.1"
marshmallow-oneofschema = "3.0.1"
marshmallow-sqlalchemy = "0.26.1"
mdit-py-plugins = "0.4.0"
mdurl = "0.1.2"
multidict = "6.0.4"
ordered-set = "4.1.0"
packaging = "21.3"
pathspec = "0.9.0"
pendulum = "2.1.2"
pluggy = "1.2.0"
prison = "0.2.1"
psutil = "5.9.5"
pycparser = "2.21"
pydantic = "1.10.11"
pygments = "2.15.1"
pyjwt = "2.7.0"
pyparsing = "3.1.0"
python-daemon = "3.0.1"
python-dateutil = "2.8.2"
python-nvd3 = "0.15.0"
python-slugify = "8.0.1"
pytz = "2023.3"
pytzdata = "2020.1"
pyyaml = "6.0"
referencing = "0.29.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rfc3986 = "1.5.0"
rich = "13.4.2"
rich-argparse = "1.2.0"
rpds-py = "0.8.8"
setproctitle = "1.3.2"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "1.4.49"
sqlalchemy-jsonfield = "1.0.1.post0"
sqlalchemy-utils = "0.41.1"
sqlparse = "0.4.4"
tabulate = "0.9.0"
tenacity = "8.2.2"
termcolor = "2.3.0"
text-unidecode = "1.3"
typing-extensions = "4.7.1"
uc-micro-py = "1.0.2"
unicodecsv = "0.14.1"
urllib3 = "1.26.16"
werkzeug = "2.2.3"
wrapt = "1.15.0"
wtforms = "3.0.1"
yarl = "1.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
